% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR.R
\name{get_IR_data}
\alias{get_IR_data}
\title{get Interest Rate}
\usage{
get_IR_data(in_ir_file, from_year = NULL)
}
\arguments{
\item{in_ir_file}{remote file or local file downloaded from http://www.freddiemac.com/pmms/docs/30yr_pmmsmnth.xls or http://www.freddiemac.com/pmms/docs/15yr_pmmsmnth.xls}

\item{from_year}{Four digits representing the data starting year.}
}
\value{

}
\description{
get Interest Rate
}
\examples{
     ir30file <- "http://www.freddiemac.com/pmms/docs/30yr_pmmsmnth.xls"
     get_IR_data(ir30file, from_year =2017)

     ir15file <- "http://www.freddiemac.com/pmms/docs/15yr_pmmsmnth.xls"
     get_IR_data(ir15file, from_year =2010)
}
