% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR.R
\name{get_IR_FRM30}
\alias{get_IR_FRM30}
\title{get FRM 30yr  interest rate}
\usage{
get_IR_FRM30(from_year = NULL)
}
\arguments{
\item{from_year}{Four digits representing the data starting year.}
}
\value{

}
\description{
get FRM 30yr  interest rate
}
\examples{
    irdata <- get_IR_FRM30()

}
