% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR.R
\name{get_IR_FRM_15_30}
\alias{get_IR_FRM_15_30}
\title{get FRM 30yr and 15yr interest rate}
\usage{
get_IR_FRM_15_30(from_year = NULL)
}
\arguments{
\item{from_year}{Four digits representing the data starting year.}
}
\value{

}
\description{
get FRM 30yr and 15yr interest rate
}
\examples{
    irdata <- get_IR_FRM_15_30()
}
