% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPI.R
\name{get_State_hpi}
\alias{get_State_hpi}
\title{get State HPI data}
\usage{
get_State_hpi(statefile, from_year = NULL, out.Monthly = TRUE)
}
\arguments{
\item{statefile}{The filename for the data downloaded from "https://www.fhfa.gov/DataTools/Downloads/Documents/HPI/HPI_AT_state.csv". If this is null, the function will connect to the download site.}

\item{from_year}{Four digits representing the data starting year.}

\item{out.Monthly}{TRUE means Monthly data, FASLE means Quarterly data.}
}
\value{
data in the data.table format with columns (State, Date, Index)
}
\description{
get State level Home Price Index data
}
\examples{
      # get recent States Home Price Index (monthly) since 2010
      statehpi_m    <- get_State_hpi(from_year=2010, out.Monthly=TRUE)
      # get recent States Home Price Index (quarterly) since 2000
      statehpi_Q    <- get_State_hpi(from_year=2000, out.Monthly=FALSE)
}
