% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPI.R
\name{get_US_hpi}
\alias{get_US_hpi}
\title{get National HPI data}
\usage{
get_US_hpi(usfile, from_year = NULL, out.Monthly = TRUE)
}
\arguments{
\item{usfile}{The filename for the data downloaded from "https://www.fhfa.gov/DataTools/Downloads/Documents/HPI/HPI_AT_us_and_census.csv". If this is null, the function will connect to the download site.}

\item{from_year}{Four digits representing the data starting year.}

\item{out.Monthly}{TRUE means Monthly data, FASLE means Quarterly data.}
}
\value{
data in the data.table format with columns (State, Date, Index)
}
\description{
get National Home Price Index data
}
\examples{
      #  get recent National (US) Home Price Index (monthly) since 2010
      ushpi_m    <- get_US_hpi(from_year=2010, out.Monthly=TRUE)
      #  get recent National (US) Home Price Index (quarterly) since 2000
      ushpi_Q    <- get_US_hpi(from_year=2000, out.Monthly=FALSE)
}
