% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPI.R
\name{get_ZIP3_hpi}
\alias{get_ZIP3_hpi}
\title{get ZIP3 HPI data}
\usage{
get_ZIP3_hpi(zip3file = NULL, from_year = NULL, out.Monthly = TRUE)
}
\arguments{
\item{zip3file}{The filename for the data downloaded from "https://www.fhfa.gov/DataTools/Downloads/Documents/HPI/HPI_AT_3zip.xlsx". If this is null, the function will connect to the download site.}

\item{from_year}{Four digits representing the data starting year.}

\item{out.Monthly}{TRUE means Monthly data, FASLE means Quarterly data.}
}
\value{
data in the data.table format with columns (ZIP3, Date, Index)
}
\description{
get ZIP3 level Home Price Index data
}
\examples{
      # get recent ZIP3 Home Price Index (monthly) since 2010
      zip3hpi_m    <- get_ZIP3_hpi(from_year=2010, out.Monthly=TRUE)
      # get recent ZIP3 (US) Home Price Index (quarterly) since 2000
      zip3hpi_Q    <- get_ZIP3_hpi(from_year=2000, out.Monthly=FALSE)
}
