% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPI.R
\name{plot_HP}
\alias{plot_HP}
\title{Plotting HPI data.}
\usage{
plot_HP(dt, regionnames = NULL, title = "Home Price Index")
}
\arguments{
\item{dt}{HPI data}

\item{regionnames}{either a list of State Names or a list of ZIP3 codes.}

\item{title}{Title appearing the chart}
}
\value{
dygraphs object
}
\description{
Plotting HPI data.
}
\examples{
     # get recent ZIP3 Home Price Index (quarterly) since 2005
     zip3hpi_Q    <- get_ZIP3_hpi(from_year=2005, out.Monthly=FALSE)
     plot_HP(dt=zip3hpi_Q,  regionnames = c("010", "011", "012") , title="Quarterly HPI for selected zips")
}
