% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR.R
\name{plot_IR}
\alias{plot_IR}
\title{Plotting IR data.}
\usage{
plot_IR(dt, title = "Interest Rate")
}
\arguments{
\item{dt}{IR data}

\item{title}{Title appearing the chart}
}
\value{
dygraphs object
}
\description{
Plotting IR data.
}
\examples{
   irdata <- get_IR_FRM_15_30()
   plot_IR(irdata, title="IR rate for FRM30yr and FRM15yr")
}
